Better organize e2e tests
